+++ src/Phan/Analysis/PostOrderAnalysisVisitor.php
@@ -2096,16 +2096,6 @@ class PostOrderAnalysisVisitor extends AnalysisVisitor
                     (string)$return_type
                 );
             }
-
-            if ($method->isStatic()
-                && $method->getUnionType()->hasTemplateType()
-            ) {
-                $this->emitIssue(
-                    Issue::TemplateTypeStaticMethod,
-                    $node->lineno ?? 0,
-                    (string)$method->getFQSEN()
-                );
-            }
         }
 
         if ($method->getHasReturn() && $method->getIsMagicAndVoid()) {