+++ src/Phan/Language/Element/Comment/ReturnComment.php
@@ -3,6 +3,7 @@ declare(strict_types=1);
 
 namespace Phan\Language\Element\Comment;
 
+use Phan\Language\Type\TemplateType;
 use Phan\Language\UnionType;
 
 /**
@@ -45,5 +46,24 @@ class ReturnComment
     public function getLineno() : int
     {
         return $this->lineno;
+    }
+
+    /**
+     * Replace the resolved reference to class T (possibly namespaced) with a regular template type.
+     *
+     * @param array<string,TemplateType> $template_fix_map maps the incorrectly resolved name to the template type
+     * @return void
+     */
+    public function convertTypesToTemplateTypes(array $template_fix_map)
+    {
+        $this->type = $this->type->withConvertTypesToTemplateTypes($template_fix_map);
+    }
+
+    /**
+     * Helper for debugging
+     */
+    public function __toString() : string
+    {
+        return "ReturnComment(type=$this->type)";
     }
 }